Windows: Factor out ulimit and refactor hostconfig
