+++ builder/internals_windows.go
@@ -1,0 +1,8 @@
+package builder
+
+import (
+	"github.com/docker/docker/runconfig"
+)
+
+func setPlatformSpecificHostConfig(b *Builder, hostConfig *runconfig.HostConfig) {
+}
