+++ runconfig/hostconfig.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/docker/docker/nat"
-	"github.com/docker/docker/pkg/ulimit"
 )
 
 type KeyValuePair struct {
@@ -38,71 +37,6 @@ func (n NetworkMode) IsNone() bool {
 	return n == "none"
 }
 
-type IpcMode string
-
-// IsPrivate indicates whether container use it's private ipc stack
-func (n IpcMode) IsPrivate() bool {
-	return !(n.IsHost() || n.IsContainer())
-}
-
-func (n IpcMode) IsHost() bool {
-	return n == "host"
-}
-
-func (n IpcMode) IsContainer() bool {
-	parts := strings.SplitN(string(n), ":", 2)
-	return len(parts) > 1 && parts[0] == "container"
-}
-
-func (n IpcMode) Valid() bool {
-	parts := strings.Split(string(n), ":")
-	switch mode := parts[0]; mode {
-	case "", "host":
-	case "container":
-		if len(parts) != 2 || parts[1] == "" {
-			return false
-		}
-	default:
-		return false
-	}
-	return true
-}
-
-func (n IpcMode) Container() string {
-	parts := strings.SplitN(string(n), ":", 2)
-	if len(parts) > 1 {
-		return parts[1]
-	}
-	return ""
-}
-
-type PidMode string
-
-// IsPrivate indicates whether container use it's private pid stack
-func (n PidMode) IsPrivate() bool {
-	return !(n.IsHost())
-}
-
-func (n PidMode) IsHost() bool {
-	return n == "host"
-}
-
-func (n PidMode) Valid() bool {
-	parts := strings.Split(string(n), ":")
-	switch mode := parts[0]; mode {
-	case "", "host":
-	default:
-		return false
-	}
-	return true
-}
-
-type DeviceMapping struct {
-	PathOnHost        string
-	PathInContainer   string
-	CgroupPermissions string
-}
-
 type RestartPolicy struct {
 	Name              string
 	MaximumRetryCount int
@@ -113,59 +47,16 @@ type LogConfig struct {
 	Config map[string]string
 }
 
-type LxcConfig struct {
-	values []KeyValuePair
-}
+// CommonHostConfig defines the host configuration which is common across
+// platforms.
+type CommonHostConfig struct {
 
-func (c *LxcConfig) MarshalJSON() ([]byte, error) {
-	if c == nil {
-		return []byte{}, nil
-	}
-	return json.Marshal(c.Slice())
-}
+	// TODO Windows. More fields may require factoring out here once the
+	// capabilities of Windows Server containers are finalised. Have kept
+	// properties which seem reasonable currently.
 
-func (c *LxcConfig) UnmarshalJSON(b []byte) error {
-	if len(b) == 0 {
-		return nil
-	}
-
-	var kv []KeyValuePair
-	if err := json.Unmarshal(b, &kv); err != nil {
-		var h map[string]string
-		if err := json.Unmarshal(b, &h); err != nil {
-			return err
-		}
-		for k, v := range h {
-			kv = append(kv, KeyValuePair{k, v})
-		}
-	}
-	c.values = kv
-
-	return nil
-}
-
-func (c *LxcConfig) Len() int {
-	if c == nil {
-		return 0
-	}
-	return len(c.values)
-}
-
-func (c *LxcConfig) Slice() []KeyValuePair {
-	if c == nil {
-		return nil
-	}
-	return c.values
-}
-
-func NewLxcConfig(values []KeyValuePair) *LxcConfig {
-	return &LxcConfig{values}
-}
-
-type HostConfig struct {
 	Binds           []string
 	ContainerIDFile string
-	LxcConf         *LxcConfig
 	Memory          int64 // Memory limit (in bytes)
 	MemorySwap      int64 // Total memory usage (memory + swap); set `-1` to disable swap
 	CpuShares       int64 // CPU shares (relative weight vs. other containers)
@@ -173,28 +64,14 @@ type HostConfig struct {
 	CpusetCpus      string // CpusetCpus 0-2, 0,1
 	CpusetMems      string // CpusetMems 0-2, 0,1
 	CpuQuota        int64
-	BlkioWeight     int64 // Block IO weight (relative weight vs. other containers)
-	OomKillDisable  bool  // Whether to disable OOM Killer or not
-	Privileged      bool
 	PortBindings    nat.PortMap
 	Links           []string
 	PublishAllPorts bool
 	Dns             []string
 	DnsSearch       []string
-	ExtraHosts      []string
-	VolumesFrom     []string
-	Devices         []DeviceMapping
 	NetworkMode     NetworkMode
-	IpcMode         IpcMode
-	PidMode         PidMode
-	CapAdd          []string
-	CapDrop         []string
 	RestartPolicy   RestartPolicy
-	SecurityOpt     []string
-	ReadonlyRootfs  bool
-	Ulimits         []*ulimit.Ulimit
 	LogConfig       LogConfig
-	CgroupParent    string // Parent cgroup.
 }
 
 func MergeConfigs(config *Config, hostConfig *HostConfig) *ContainerConfigWrapper {
