+++ runconfig/hostconfig_unsupported.go
@@ -1,0 +1,7 @@
+// +build !windows,!linux
+
+package runconfig
+
+type HostConfig struct {
+	CommonHostConfig
+}
