+++ runconfig/hostconfig_windows.go
@@ -1,0 +1,10 @@
+// +build windows
+
+package runconfig
+
+type HostConfig struct {
+	CommonHostConfig
+
+	// Fields below here are platform specific. There are none currently
+	// for the Windows platform.
+}
